Stab at graphiti. Decent front end for graphite.
This pulls in a few package requirements, clones the repo and installs
the gems via bundler, then adds it via nginx/unicorn modules. It works
magic via exec{}s to gem install using ruby 1.9 which is a requirement of
graphiti. It pulls in unstable to grab a version of redis-server that
it needs too (bpo version fails). It has a commented out amazon_s3.yml due
to me not wanting to push Amazon credentials to git.

It also makes a new init script for unicorn, to use Ruby 1.9, this
could/should migrate to a heavily defaulted to 1.8 template at some stage.