Add rbenv detection and support.
When starting the `${name}_user` variable will be checked for the
presence of a `~${user}/.rbenv` directory. If this directory is present
then the `.rbenv/shims` and `.rbenv/bin` directories will be prepended
to the PATH so that rbenv can be used to load unicorn.

This support primarily requires the use of a binstub for unicorn to be
created and for the path to this binstub to be given as the `source`
parameter of the `unicorn::app` resource.